refactor(pricing): extract request-pricing arithmetic into its own module
The cost maths for a single request — list price, the cache-adjusted charged
cost, the cache-split safety rules and as-of tariff resolution — lived inside
the analytics batcher, reachable only from the live write path.

An upcoming usage-recompute feature has to re-price historical requests when a
bug wrote the wrong token counts. If it priced them by its own copy of this
arithmetic, a "correction" could itself be wrong — the two implementations would
drift silently and only show up on a customer's invoice. So the maths moves out
to be shared rather than duplicated.

Behaviour-preserving. The functions now take a narrow `TokenCounts` instead of a
`RawAnalyticsRecord`, so neither caller owns the other's record type, and the
error component is a parameter so a recompute anomaly is distinguishable from a
live-path one on dashboards instead of both reporting as the batcher.

Two things worth noting for review:

- `find_best_tariff` was a `&self` method that never touched `self`, and the
  tests had a hand-copied reimplementation of it to avoid constructing a
  batcher. It is now a free function and the tests exercise the real one — the
  duplicate is gone.
- Test coverage moves with the code and is otherwise unchanged. What stays in
  the batcher is what the batcher still owns: the `RawAnalyticsRecord` →
  `TokenCounts` field mapping (a transposed cache tier there would bill reads at
  a write multiplier) and its local list-price wrapper.
